requirements.txt
Django
Pillow
instagram-scraper
selenium
django_rest_framework
